chore(rename): xyz.mobilestack -> org.tucop.wallet (tier 3)
Renames the Android internal Java package from the legacy Valora/
MobileStack lineage to org.tucop.wallet, aligned with reverse-DNS of
the app's applicationId (org.tucop) and leaving room for future TuCop
apps (org.tucop.ramp, org.tucop.backend).

Not user-visible. Play Store install continuity preserved:
applicationId stays org.tucop, keystore alias stays mobilestack-key-
alias in gradle.properties, bundle ID stays org.tucop. This only
changes the Java/Kotlin source package + AGP namespace + BuildConfig
package, which show up in Sentry stack traces and Android Studio logs.

Changes:
- android/app/src/main/java/xyz/mobilestack/ -> android/app/src/main/java/org/tucop/wallet/
  (5 files: MainActivity.kt, MainApplication.kt, FirebaseMessagingService.java,
  UserAgentClientFactory.java, UserAgentInterceptor.java)
- android/app/src/androidTest/java/xyz/mobilestack/DetoxTest.java ->
  android/app/src/androidTest/java/org/tucop/wallet/DetoxTest.java
- All 'package xyz.mobilestack' declarations updated to 'package org.tucop.wallet'
- DetoxTest.java fully-qualified 'xyz.mobilestack.BuildConfig.DEBUG' updated
- android/app/build.gradle: namespace + resValue build_config_package
- android/app/proguard-rules.pro: -keep class rule updated
- android/app/src/androidTest/AndroidManifest.xml: package attr updated

Verification:
- ./gradlew :app:assembleMainnetDebug: BUILD SUCCESSFUL in 4m 27s.
- aapt2 dump packagename on the built APK returns 'org.tucop' (applicationId
  intact, Play Store install continuity preserved).
- yarn build:ts clean.
- yarn lint clean.
