+++ android/app/build.gradle
@@ -141,7 +141,7 @@ def secrets = loadProjectSecrets("mainnet")
 
 android {
     compileSdkVersion rootProject.ext.compileSdkVersion
-    namespace "xyz.mobilestack"
+    namespace "org.tucop.wallet"
 
     // AGP 8.x configuration - removed obsolete aaptOptions
 
@@ -172,7 +172,7 @@ android {
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         resValue "string", "app_name", project.env.get("APP_DISPLAY_NAME")
         resValue "string", "deep_link_url_scheme", project.env.get("DEEP_LINK_URL_SCHEME")
-        resValue "string", "build_config_package", "xyz.mobilestack"
+        resValue "string", "build_config_package", "org.tucop.wallet"
         missingDimensionStrategy 'react-native-camera', 'general'
         vectorDrawables.useSupportLibrary = true
         resValue "bool", "is_profiling_build", System.getenv("IS_PROFILING_BUILD") ?: "false"
