+++ android/app/proguard-rules.pro
@@ -20,7 +20,7 @@
 -keep class com.segment.analytics.** { *; }
 -keep class androidx.lifecycle.DefaultLifecycleObserver
 
--keep class xyz.mobilestack.BuildConfig { *; }
+-keep class org.tucop.wallet.BuildConfig { *; }
 -keep public class com.horcrux.svg.** {*;}
 -keep class com.rt2zz.reactnativecontacts.** {*;}
 -keepclassmembers class com.rt2zz.reactnativecontacts.** {*;}
