+++ android/app/src/main/java/org/tucop/wallet/UserAgentInterceptor.java
@@ -1,4 +1,4 @@
-package xyz.mobilestack;
+package org.tucop.wallet;
 
 import android.os.Build;
 import java.io.IOException;
